Dashboard For Discharges
Discharge Dashboard tracks the number of discharges with respect year of beginning date., payer source. Tracks the Change in percentage of discharges based on the payer source over year and also evaluates the change in number of discharges by every year with respect to the previous year and labeled by the payer source.